+++ relay-event-schema/src/protocol/contexts/mod.rs
@@ -17,6 +17,7 @@ mod reprocessing;
 mod response;
 mod runtime;
 mod spring;
+mod threadpool_info;
 mod trace;
 mod unity;
 mod user_report_v2;
@@ -38,6 +39,7 @@ pub use reprocessing::*;
 pub use response::*;
 pub use runtime::*;
 pub use spring::*;
+pub use threadpool_info::*;
 pub use trace::*;
 pub use unity::*;
 pub use user_report_v2::*;
@@ -100,6 +102,9 @@ pub enum Context {
     PerformanceScore(Box<PerformanceScoreContext>),
     /// Spring / Spring Boot information.
     Spring(Box<SpringContext>),
+    /// Thread pool information.
+    #[metastructure(tag = "threadpool_info")]
+    ThreadPoolInfo(Box<ThreadPoolInfoContext>),
     /// OTA Updates information.
     OTAUpdates(Box<OTAUpdatesContext>),
     /// Chromium Stability Report from minidump.